fix fabric + some final refactoring
