+++ forge/src/main/java/org/valkyrienskies/mod/forge/mixin/feature/crash_report/CrashReportExtenderMixin.java
@@ -1,4 +1,4 @@
-package org.valkyrienskies.mod.forge.mixin;
+package org.valkyrienskies.mod.forge.mixin.feature.crash_report;
 
 import net.minecraft.CrashReport;
 import net.minecraftforge.logging.CrashReportExtender;
@@ -6,14 +6,13 @@ import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
-import org.valkyrienskies.mod.common.VSCrashReportHeader;
+import org.valkyrienskies.mod.common.crash_report.VSCrashReportHeader;
 
 @Mixin(CrashReportExtender.class)
 public class CrashReportExtenderMixin {
 
     @Inject(method = "addCrashReportHeader", at = @At("HEAD"), remap = false)
-    private static void addConnectorCrashReportHeader(StringBuilder stringbuilder, CrashReport crashReport, CallbackInfo ci) {
+    private static void addCrashReportHeader(StringBuilder stringbuilder, CrashReport crashReport, CallbackInfo ci) {
         VSCrashReportHeader.addCrashReportHeader(stringbuilder);
     }
-
 }
