+++ forge/src/main/resources/valkyrienskies-forge.mixins.json
@@ -25,6 +25,7 @@
     "entity.MixinBoat",
     "entity.MixinContainerEntity",
     "feature.air_pockets.MixinEntityFluidPushing",
+    "feature.crash_report.CrashReportExtenderMixin",
     "feature.fix_tp_ships_cross_dimension.MixinLevelChunk",
     "feature.forge_interact.MixinIForge2Player",
     "feature.forge_interact.MixinIForgePlayer",
@@ -36,8 +37,7 @@
     "feature.mob_spawning.MixinPhantomSpawnerShipFinalize",
     "feature.shipyard_entities.MixinPersistentEntitySectionManager",
     "feature.water_in_ships_entity.MixinEntity",
-    "world.level.block.FireMixin",
-    "CrashReportExtenderMixin"
+    "world.level.block.FireMixin"
   ],
   "client": [
     "client.render.MixinLevelRenderer",
